rity: config: verify udev rules with md5 sum
Checking the presence of the udev rules is not enough because as
the udev rules evolve, rity-config will always returns that the
file is good despite being outdated. Use the MD5 of the files in
order to easily be able to check whether we are using the latest
version of the udev rules.

Signed-off-by: Fabien Parent <fparent@baylibre.com>