rity: image: yocto: allow flashing WIC files
This allow flashing the whole WIC file instead of flashing its
different components separately.

In order to be able to flash the WIC, we need to have a sparse image
otherwise fastboot fails to flash the image. In order to do so, we
use img2simg python project.

The sparse image will be generated at time of flashing.

Signed-off-by: Fabien Parent <fparent@baylibre.com>