image: android: don't use --uboot-env-size arg
When looking at the help from rity-flash, we can see the following
commands in the Android section:

  Android:
    --dtbo-index DTBO_INDEX
                          Enable one or multiple DTBO(s)

However, in AndroidImage, we use self self.args.uboot_env_size, which is
an undocumented argument.

It works, because we are lucky:
- YoctoImage gets parsed before AndroidImage
- YoctoImage adds the --uboot-env-size argument to the common argparser class
- AndroidImage gets parsed, and can access args.uboot_env_size

Let's remove this, as it's confusing.
Default to a hard-coded (identical) value instead.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>