image: android: add --serialno option
For automation purposes, it's convenient to have different serial
numbers. Those serial numbers are used when enumerating a board in
fastboot or using adb.

This serial number is stored in a U-Boot environment variable (serial#)

Add a --serialno option which allows us to patch the
u-boot-env with a custom serial number.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>