flash: Check image content and fail early
Sometimes we need to make the board entering download mode to see error
messages of invalid image content, which is not intuitive in terms of
UX. If something wrong in an image, the tool should report error and
exit earlier.

Thus we add new method Flash.check() to examine current image content
and command line options before starting flashing process.

Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>