image: ubuntu: Add uboot_env_size and uboot_env_offset_redund config
Here is the sample in aiot-tools config file for new values:
ubuntu.json:
"uboot-env": {
  "env-size": "8192",
  "env-offset-redund": "0x2000"
}

In order to support both Ubuntu and Yocto images with a single U-Boot
defconfig, some settings for Ubuntu have to align with Yocto. Modifying
the default value directly will lead to the backward compatibility
issue. So, keeping the default values and make them configurable for
backward compatibility.