flash: Generalize uboot-env-redund-offset parameter
Redundant U-Boot environment data is required by Ubuntu Core images,
and also Yocto images in the future if the support for EFI boot is to be added.

Therefore we add a common parameter "--uboot-env-redund-offset".
If this is defined, aiot-flash would generate a redundant copy
of u-boot-env and write it in the given offset.

Since the size of U-boot environment might also need to be adjusted,
we also move "--uboot-env-size" to common parameter.

The setting in `ubuntu.json` in Ubuntu images types take priority
over this commandline parameter. A warning message would appear
if both the cmd param and the json settings are assigned.

Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>