ubootenv: Accept key-value argument to update uboot env
Add an argument --uboot-env-set (-e) to allow
updating or adding U-Boot env variables using one or more KEY=VALUE pair.

Switching to ebbr boot target and set Ethernet MAC address:

$ genio-flash -e boot_targets=ebbr ethaddr=00:11:22:33:44:55

This feature does not verify if the value is a semantically correct
U-Boot environment variable, it simply update the variable as-is.

Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>