image: yocto: Add dtbo list from u-boot-initial-env
Fix issue that the dtbo enabled by KERNEL_DEVICETREE_OVERLAYS_AUTOLOAD
in Yocto local.conf does not appear in the reported list of dtbo,
for example:

```
$ genio-flash
AIoT Tools: v1.3.2.dev0
Yocto Image:
        name:     Rity Demo Image (rity-demo-image)
        distro:   Rity Demo Layer 23.1-dev (rity-demo)
        codename: kirkstone
        machine:  genio-700-evk
        overlays: []
```

The "overlays" list is empty even if "video.dtbo" and "gpu-mali.dtbo"
are loaded by default.

If user assigns '--loadt-dtbo a.dtbo' arguments, the new 'a.dtbo' is
add on top of the ones defined in u-boot-initial-env.

CR-Id: AUTO00249839
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>