bootrom: Launch aiot-bootrom as a module
genio-bootroom is a install dependency of genio-tools.
However, bootrom.py that run it is using subprocess,
which assume genio-bootroom to be available in a system-wide location.

This breaks genio-tools in virtualenv context. So,
run genio-bootrom as a Python module instead of as a script.

Also update genio-bootrom version dependencies to 1.1.11 or later,
so we could run it as a module.

Bug: https://bugs.launchpad.net/baoshan/+bug/2081730
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>