bootfirmware: add support for rity-bootfirmware-image (eMMC)
This commit introduces support for the rity-bootfirmware-image, a
configuration designed to enhance the boot process by providing compatibility
with EFI boot and ARM SystemReady standards. Developers can now build a boot
firmware-only image (without kernel and rootfs) and flash it to the device.
This allows developers to install Linux distro CD images via a USB disk with
this boot firmware image.

Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>