image: Add Raw image type
Add a raw image type to support simple disk images
for eMMC and UFS.

The primary usage of this is to support raw disk images
converted from MediaTek SP Flash Tool using an internal
tool.

If a `raw_image.json` is given, use the json file
as partition file setting.

If not given, this image type checks a default set
of image binary names. The name came from the
internal conversion tool.

Refer to the document named
"Converting Android Image to RAW Image and Flashing Using Public Tool"
for detail.

The following operations are supported:

- genio-flash erase-mmc
- genio-flash

Note that U-Boot commands are not explictly blocked, but the flash
process simply skips the partition image generation logic for
mmc0boot1.

Bug: IOTXYCTO-23
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>