multi_download_cli: fix a error caused by flash daemon not ready
Because the genio-flash daemon has undergone some refactoring,
when multiple download command-line client are running in other
consoles, exceptions of reading 'status_info' for each worker
may occur if the flash daemon is not ready.
In this case, simply skip the exception so the client continues
waiting for the daemon to become available.

Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>