flash: remove partitions check for erase* actions
When we execute erase or erase_after_flash actions, the partition may
not be present in img.partitions table.
Indeed for Android we need to erase misc partition but we never flash
a misc image.
For these actions we don't need to check if the partition is present
in the img.partitions table.

Signed-off-by: Julien Masson <jmasson@baylibre.com>